requirements.txt
django==1.6.1
south==0.8.4
psycopg2==2.5.2
-e git+git://github.com/newsapps/django-boundaryservice.git@v0.2.2#egg=django-boundaryservice
djangorestframework==2.3.13
djangorestframework-gis==0.2
Jinja2==2.7.2
jingo==0.7
django-mptt==0.6.0
jsonfield==0.9.20
django-extensions==1.3.3
mock==1.0.1
django-nose==1.2
nose==1.3.1
dj-database-url==0.3.0
ipython
ipdb
ipdbplugin==1.4
numpy
scipy
Markdown==2.4.1
django-cors-headers==0.12
